+++ Modules/Intent.Modules.EntityFrameworkCore/Templates/EntityTypeConfiguration/EntityTypeConfigurationTemplatePartial.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using System.Reflection.Metadata;
 using System.Text;
 using System.Threading;
 using Intent.Engine;
@@ -11,6 +12,7 @@ using Intent.Metadata.RDBMS.Api;
 using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
+using Intent.Modules.Common.CSharp.Mapping;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.CSharp.VisualStudio;
 using Intent.Modules.Common.Templates;
@@ -48,6 +50,7 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.EntityTypeConfiguration
             AddNugetDependency(NugetPackages.MicrosoftEntityFrameworkCore(Project));
             AddTypeSource("Domain.Entity");
             AddTypeSource("Domain.ValueObject");
+            AddTypeSource("Intent.Entities.DomainEnum");
 
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .AddUsing("Microsoft.EntityFrameworkCore")
@@ -71,7 +74,8 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.EntityTypeConfiguration
                             }
 
                             method.AddStatements(GetTypeConfiguration(Model.InternalElement, @class, null));
-                            method.AddStatements(GetCheckConstraints(Model));
+                            //Moved this into GetTypeConfiguration as its applicabale for Owned Tables too
+                            //method.AddStatements(GetCheckConstraints(Model));
                             method.Statements.SeparateAll();
 
                             AddIgnoreForNonPersistent(method, isOwned: false);
@@ -236,6 +240,11 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.EntityTypeConfiguration
             statements.AddRange(GetAssociations(targetType)
                 .Where(RequiresConfiguration)
                 .Select(x => GetAssociationMapping(x, targetType, @class)));
+
+            if (targetType.IsClassModel())
+            {
+                statements.AddRange(GetCheckConstraints(targetType.AsClassModel()).Select(s => new CSharpStatement(s)));
+            }
 
             return statements.Where(x => x != null).ToList();
         }
@@ -544,11 +553,42 @@ namespace Intent.Modules.EntityFrameworkCore.Templates.EntityTypeConfiguration
 
         private IEnumerable<string> GetCheckConstraints(ClassModel model)
         {
+            if (ExecutionContext.Settings.GetDatabaseSettings().EnumCheckConstraints() && model.Attributes.Any(a => a.TypeReference.Element.IsEnumModel()))
+            {
+                var enumsStoredAsStrings = ExecutionContext.Settings.GetDatabaseSettings().StoreEnumsAsStrings();
+                var attributesToAdd = model.Attributes.Where(a => a.TypeReference.Element.IsEnumModel());
+                foreach (var enumAttribute in attributesToAdd)
+                {
+                    AddUsing("System");
+                    AddUsing("System.Linq");
+                    var constraint = new StringBuilder();
+                    string enumType = this.GetTypeName(enumAttribute.TypeReference);
+                    if (ExecutionContext.Settings.GetDatabaseSettings().StoreEnumsAsStrings())
+                    {
+                        constraint.AppendLine( @$"var enumValues =  Enum.GetNames<{enumType}>()
+                .Select(e => $""'{{e}}'"");");
+                    }
+                    else
+                    {
+                        constraint.AppendLine(@$"var enumValues = Enum.GetValuesAsUnderlyingType<{enumType}>()
+                .Cast<object>()
+                .Select(value => value.ToString());");
+                    }
+                    constraint.AppendLine();
+                    constraint.AppendLine(@$"builder.ToTable(tb => tb.HasCheckConstraint(""{GetContraintName(enumAttribute)}"", $""\""{enumAttribute.Name}\"" IN ({{string.Join("","", enumValues)}})""));");
+                    yield return constraint.ToString();
+                }
+            }
             var checkConstraints = model.GetCheckConstraints();
             foreach (var checkConstraint in checkConstraints)
             {
                 yield return @$"builder.HasCheckConstraint(""{checkConstraint.Name()}"", ""{checkConstraint.SQL()}"");";
             }
+        }
+
+        private string GetContraintName(AttributeModel Model)
+        {
+            return (Model.InternalElement.ParentElement.Name + Model.Name + "Check").ToSnakeCase();
         }
 
         private CSharpStatement[] GetIndexes(ClassModel model)