+++ Modules/Intent.Modules.Security.JWT/Decorators/ConfigurationSettingsJWTDecorator.cs
@@ -8,8 +8,8 @@ using Intent.Modules.Common.Templates;
 using Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings;
 using Intent.RoslynWeaver.Attributes;
 
-[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecorator", Version = "1.0")]
 [assembly: DefaultIntentManaged(Mode.Merge)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecorator", Version = "1.0")]
 
 namespace Intent.Modules.Security.JWT.Decorators
 {