+++ Modules/Intent.Modules.Security.JWT/Decorators/ConfigurationSettingsJWTDecoratorRegistration.cs
@@ -8,8 +8,8 @@ using Intent.Modules.Common.Registrations;
 using Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings;
 using Intent.RoslynWeaver.Attributes;
 
-[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecoratorRegistration", Version = "1.0")]
 [assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecoratorRegistration", Version = "1.0")]
 
 namespace Intent.Modules.Security.JWT.Decorators
 {
@@ -17,6 +17,7 @@ namespace Intent.Modules.Security.JWT.Decorators
     [Description(ConfigurationSettingsJWTDecorator.DecoratorId)]
     public class ConfigurationSettingsJWTDecoratorRegistration : DecoratorRegistration<AppSettingsTemplate, AppSettingsDecorator>
     {
+        [IntentManaged(Mode.Fully)]
         public override AppSettingsDecorator CreateDecoratorInstance(AppSettingsTemplate template, IApplication application)
         {
             return new ConfigurationSettingsJWTDecorator(template, application);