+++ Tests/OutputCachingRedis.Tests/OutputCachingRedis.Tests.Application/Accounts/UpdateAccount/UpdateAccountCommand.cs
@@ -1,0 +1,22 @@
+using System;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+using OutputCachingRedis.Tests.Application.Common.Interfaces;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandModels", Version = "1.0")]
+
+namespace OutputCachingRedis.Tests.Application.Accounts.UpdateAccount
+{
+    public class UpdateAccountCommand : IRequest, ICommand
+    {
+        public UpdateAccountCommand(string name, Guid id)
+        {
+            Name = name;
+            Id = id;
+        }
+
+        public string Name { get; set; }
+        public Guid Id { get; set; }
+    }
+}