+++ Tests/OutputCachingRedis.Tests/OutputCachingRedis.Tests.Application/Accounts/UpdateAccount/UpdateAccountCommandValidator.cs
@@ -1,0 +1,24 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "2.0")]
+
+namespace OutputCachingRedis.Tests.Application.Accounts.UpdateAccount
+{
+    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    public class UpdateAccountCommandValidator : AbstractValidator<UpdateAccountCommand>
+    {
+        [IntentManaged(Mode.Merge)]
+        public UpdateAccountCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        private void ConfigureValidationRules()
+        {
+            RuleFor(v => v.Name)
+                .NotNull();
+        }
+    }
+}