+++ Tests/OutputCachingRedis.Tests/OutputCachingRedis.Tests.Domain/Repositories/IAccountRepository.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using OutputCachingRedis.Tests.Domain.Entities;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Entities.Repositories.Api.EntityRepositoryInterface", Version = "1.0")]
+
+namespace OutputCachingRedis.Tests.Domain.Repositories
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public interface IAccountRepository : IEFRepository<Account, Account>
+    {
+        [IntentManaged(Mode.Fully)]
+        Task<Account?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
+        [IntentManaged(Mode.Fully)]
+        Task<List<Account>> FindByIdsAsync(Guid[] ids, CancellationToken cancellationToken = default);
+    }
+}