+++ Modules/Intent.Modules.AspNetCore.OutputCaching.Redis/release-notes.md
@@ -1,5 +1,11 @@
-### Version 1.0.2
+### Version 1.1.1
 
+- Improvement: `InstanceName` configuration setting will by default take on the Application name.
+
+### Version 1.1.0
+
+- Improvement: Added `Caching Policies` element to centralize `Caching Policy`s.
+- Improvement: General better out the box experiences.
 - Improvement: Updated NuGet packages to latest stables.
 
 ### Version 1.0.1